requirements.txt
python-telegram-bot==20.7
gdown
python-dotenv
